Fixed Linux Build

+++ extern/raknet/include/raknet/BitStream.h
@@ -1682,7 +1682,7 @@ namespace RakNet
 			y=0.0;
 		else
 		{
-			ReadCompressed((float)cy);
+			ReadCompressed(cy);
 			y=cy;
 			//Read(sy);
 			//y=((float)sy / 32767.5f - 1.0f);
@@ -1699,7 +1699,7 @@ namespace RakNet
 			//		return false;
 
 			//	z=((float)sz / 32767.5f - 1.0f);
-			if (!ReadCompressed((float)cz))
+			if (!ReadCompressed(cz))
 				return false;
 			z=cz;
 		}
